Implemented User Choice 3 and fixed getCharacter
User Choice 3 updates a character of the user's choice and allows them
to change the character's name, HP, and power. The character will then
be updated with this new information. I also changed getCharacter to
search through the array to find the character and return it. The
JavaDoc for this method was also updated.